#525 Adjust design to new style of Nextcloud app
Move branding logic to BrandingUtil
