+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/PushNotificationActivity.java
@@ -24,6 +24,7 @@ import it.niedermann.nextcloud.deck.ui.exception.ExceptionHandler;
 
 import static android.graphics.Color.parseColor;
 import static it.niedermann.nextcloud.deck.persistence.sync.adapters.db.util.LiveDataHelper.observeOnce;
+import static it.niedermann.nextcloud.deck.ui.branding.BrandingUtil.getSecondaryForegroundColorDependingOnTheme;
 
 public class PushNotificationActivity extends BrandedActivity {
