+++ app/src/main/res/layout/item_card.xml
@@ -146,6 +146,7 @@
                     android:background="?attr/selectableItemBackgroundBorderless"
                     android:contentDescription="@string/label_menu"
                     android:padding="@dimen/spacer_1hx"
+                    android:tint="@color/accent"
                     app:srcCompat="@drawable/ic_menu" />
             </LinearLayout>
         </LinearLayout>
