+++ app/src/main/res/values/styles.xml
@@ -5,6 +5,7 @@
         <item name="colorPrimaryDark">@color/primary</item>
         <item name="colorAccent">@color/accent</item>
         <item name="toolbarStyle">@style/toolbarStyle</item>
+        <item name="android:windowBackground">?attr/colorPrimary</item>
     </style>
 
     <style name="toolbarStyle" parent="@style/Widget.AppCompat.Toolbar">
